+++ picturebot/poco.py
@@ -14,6 +14,24 @@ class Config:
     Editing: str = ""
     Instagram: str = ""
 
+    def __str__(self):
+        '''Overide the __str__ class
+        
+        Returns:
+            (string): Returns the overrode string
+        '''
+        
+        return f"""
+                Workspace: {self.Workspace}
+                Workflow: {self.Workflow}
+                Baseflow: {self.Baseflow}
+                Backup: {self.Backup}
+                Selection: {self.Selection}
+                Edited: {self.Edited}
+                Preview: {self.Preview}
+                Editing: {self.Editing}
+                Instagram: {self.Instagram}
+                """
 
 @dataclass
 class Context: