+++ picturebot/shoot.py
@@ -14,35 +14,49 @@ import picturebot.base as baseflow
 import picturebot.flow as otherflow
 
 class Shoot:
-    def __init__(self, context, name):
+        
+    def __init__(self, context, index, name):
+        '''Shoot class constructor
+        
+        Args:
+            context (object): Global context object
+            index (int): Index of the selected workspace configuration
+            name (string): Shootname containing the date
+        '''        
+        
+        self.index = int(index)
         self.ctx = context
         self.name = name
     
     def Create(self):
-        # Obtain the path to the base flow project
-        pathToBaseflowProject = helper.FullFilePath(self.ctx.Config.Workspace, self.name)
+        '''Creates a shoot within the workspace directory'''
 
-        # Check whether the the path to the base flow project exists
-        if not grd.Filesystem.IsPath(pathToBaseflowProject):
-            directory.CreateFolder(pathToBaseflowProject)
-            grd.Filesystem.PathExist(pathToBaseflowProject)
+        # Obtain the root directory of the new shoot
+        shootRootPath = helper.FullFilePath(self.ctx.Config[self.index].Workspace, self.name)
+
+        # Check whether the root directory of the new shoot exists
+        if not grd.Filesystem.IsPath(shootRootPath):
+            # Create a new shoot
+            directory.CreateFolder(shootRootPath)
+            # Check whether the shoot is successfully created
+            grd.Filesystem.PathExist(shootRootPath)
         else:
-            print('Path already exsists')
+            print('Path already exists')
 
-        self.__CreateFlow(pathToBaseflowProject)
+        self.__CreateFlow(shootRootPath)
 
-    def __CreateFlow(self, name):
-        ''' Create a new flows
+    def __CreateFlow(self, root):
+        ''' Create all flows configured in the config file
 
         Args:
-            context (object): Global context object
+            root (string): Root directory of the newly created shoot
         '''
 
         counter = 0
 
         #Loop-over the workflows
-        for flow in self.ctx.Config.Workflow:
-            pathToFlow = helper.FullFilePath(name, flow)
+        for flow in self.ctx.Config[self.Index].Workflow:
+            pathToFlow = helper.FullFilePath(root, flow)
 
             # Only create non existing flows
             if not grd.Filesystem.IsPath(pathToFlow):